Move RemoteBackupListener class to separate file
